refactor(qti): extract shared vnode/shuffle utils across interactions
